transport: Track client state for real
Right now when the client state changes the respective update is
performed on the system.clients table. While doing it some bits
from this state are lost from the in-memory structures. For the
sake of exporting this information we need to track whether the
connected client goes authenticating or is already ready.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>