flat_mutation_reader: fix increasing-partition-key sanity check
The flat mutation reader has a sanity check that partition keys are
increasing. This check in is stricter in debug mode - it checks that
the seen partition keys are *strictly* increasing (this stricter check
is slower, as it involves comparing the full keys, not just the tokens).

For the very first partition key, the check should obviously succeed
because nothing had been seen previously. The way this was implemented
was to initialize _prev_partition_key to an empty partition key, which
is known to come before all keys. However, when we allow empty partition
keys to be used in the data (as we plan to do in the next patch), this
can confuse the strictly-increasing-check code, making it think it saw
the empty partition key twice...

So in this patch we change _prev_partition_key to be an std::optional.
It is initially unset, and only set when the first key is seen. The
first key can very well be an empty key without triggering an error.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>