commitlog: Keep track of last flush callback position
If we flush size-based (try to remove X bytes), we should not restrict
ourselves to bytes already sent flushes for, since that might not help
at this point. Keep track of highest flush-callback position and
restart counting from where ever this was in list.

Since RP:s grow monotonicall, we know later elements in
segment list are always bigger